refactor trello api
